make mast float

+++ solvers/denoiser.py
@@ -89,7 +89,7 @@ def recon_loss(noise, loc, scale):
             noisy256 = x_noisy * 255
             gen256 = x_gen * 255
 
-            mask = (gen256 > 0).detach().clone()
+            mask = (gen256 > 0).detach().clone().float()
             mask.requires_grad = False
 
             valid_gen256 = gen256 * mask
@@ -105,7 +105,7 @@ def recon_loss(noise, loc, scale):
             # nll = gen256 - noisy256 * torch.log(gen256 + 1e-10)
             # print(torch.log(gen256 + 1e-10).mean().item(), gen256.min().item(), '\n')
             return nll.view(len(x_noisy), -1).sum(dim=1).mean() / 255
-        
+
     elif noise == 'logistic':
         def _recon(x_gen, x_noisy):
             delta = x_noisy - x_gen
